cron to automatically scan & mount glusterfs disk. grep -c instead of wc -l

+++ glusterfs/client-setup.sh
@@ -11,9 +11,14 @@ if [ "${exit_code}" -ne 0 ]; then
     exit $exit_code
 fi
 
-# update fstab, remove the same previous entry if exists
-temp_file=$(mktemp)
-grep -vF "${glusterfs_master_host}:/gv-chia" /etc/fstab | tee "${temp_file}"
-echo "${glusterfs_master_host}:/gv-chia ${HOME}/gv-chia glusterfs defaults,_netdev,x-systemd.requires=glusterd.service,x-systemd.automount 0 0" | tee -a "${temp_file}" > /dev/null
-sudo tee /etc/fstab > /dev/null < "${temp_file}"
-rm -f "${temp_file}"
+fstab_count=$(grep -c "${glusterfs_master_host}:/gv-chia" /etc/fstab)
+if [ "${fstab_count}" -gt 0 ]; then
+  # update fstab, remove the same previous entry if exists
+  temp_file=$(mktemp)
+  grep -vF "${glusterfs_master_host}:/gv-chia" /etc/fstab | tee "${temp_file}"
+  echo "${glusterfs_master_host}:/gv-chia ${HOME}/gv-chia glusterfs defaults,_netdev,x-systemd.requires=glusterd.service,x-systemd.automount 0 0" | tee -a "${temp_file}" > /dev/null
+  sudo tee /etc/fstab > /dev/null < "${temp_file}"
+  rm -f "${temp_file}"
+else
+  echo "${glusterfs_master_host}:/gv-chia ${HOME}/gv-chia glusterfs defaults,_netdev,x-systemd.requires=glusterd.service,x-systemd.automount 0 0" | sudo tee -a /etc/fstab > /dev/null
+fi
